+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+#2015-11-10
+ - Allow complex expressions in breadcrumb names and route parameters
+
 #2014-04-18
  - Rewrite documentation